+++ NEWS.md
@@ -10,9 +10,11 @@ Penalized Regression with Second-Generation P-Values
 
 ## ProSGPV 0.2.0
 
-3/30/2021: ProSGPV can handle GLM now!
+3/31/2021: ProSGPV can handle GLM now!
 
 * Added functionalities of variable selection in Logistic, Poisson, and Cox models
+
+* Added the `which.sgpv` function to find the most frequent model when the two-stage algorithm is used
 
 * Added the S3 `plot` method for the one-stage algorithm